+++ transport/quic/quic.go
@@ -229,6 +229,7 @@ func (t *Transport) Negotiate(ctx context.Context, conn net.Conn, role transport
 		if err != nil {
 			return peer, err
 		}
+		hs = common.MergeHandshake(hs, peer)
 		if err := common.ValidateHandshake(hs, peer); err != nil {
 			return peer, err
 		}
@@ -242,6 +243,7 @@ func (t *Transport) Negotiate(ctx context.Context, conn net.Conn, role transport
 		if err != nil {
 			return peer, err
 		}
+		hs = common.MergeHandshake(hs, peer)
 		if err := common.ValidateHandshake(hs, peer); err != nil {
 			return peer, err
 		}