+++ transport/tcp_tls/tcp_tls.go
@@ -256,6 +256,7 @@ func (t *Transport) Negotiate(ctx context.Context, conn net.Conn, role transport
 		if err != nil {
 			return peer, err
 		}
+		hs = common.MergeHandshake(hs, peer)
 		if err := common.ValidateHandshake(hs, peer); err != nil {
 			return peer, err
 		}
@@ -269,6 +270,7 @@ func (t *Transport) Negotiate(ctx context.Context, conn net.Conn, role transport
 		if err != nil {
 			return peer, err
 		}
+		hs = common.MergeHandshake(hs, peer)
 		if err := common.ValidateHandshake(hs, peer); err != nil {
 			return peer, err
 		}